+++ custom-metronome/src/partials/Branding/Branding.scss
@@ -1,0 +1,34 @@
+.branding-container {
+    margin-top: 64px;
+    display: grid;
+    grid-template-columns: 40% 60%;
+    align-items: center;
+    padding: 0 64px;
+    color: #000;
+
+    .branding-call {
+        h1 {
+            font-size: 60px;
+            font-weight: 400;
+        }
+
+        h3{
+            margin-top: 32px;
+            font-size: 20px;
+            font-weight: 200;
+        }
+    }
+
+    .branding-actions{
+        display: grid;
+        margin-top: 32px;
+    }
+
+    .branding-image {
+        img {
+            width: 100%;
+            height: 100%;
+            object-fit: cover;
+        }
+    }
+}
