+++ custom-metronome/src/shared/services/button-color-creator.service.ts
@@ -1,0 +1,30 @@
+import { createTheme, Theme } from '@mui/material/styles';
+
+const newTheme: Theme = createTheme({
+    palette: {
+        neutral: {
+            main: '#64748B',
+            contrastText: '#fff',
+        },
+    },
+});
+
+declare module '@mui/material/styles' {
+    interface Palette {
+        neutral: Palette['primary'];
+    }
+
+    // allow configuration using `createTheme`
+    interface PaletteOptions {
+        neutral?: PaletteOptions['primary'];
+    }
+}
+
+// Update the Button's color prop options
+declare module '@mui/material/Button' {
+    interface ButtonPropsColorOverrides {
+        neutral: true;
+    }
+}
+
+export default newTheme;
